pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-guardian"
version = "0.1.0"
description = "Self-healing documentation for every project"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Anthropic, PBC", email = "support@anthropic.com"}
]
maintainers = [
    {name = "Anthropic, PBC", email = "support@anthropic.com"}
]
keywords = [
    "documentation",
    "docs",
    "maintenance",
    "healing",
    "markdown",
    "links",
    "automation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
requires-python = ">=3.8"
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml>=5.4"]
jinja = ["jinja2>=3.0"]
toml = ["toml>=0.10; python_version < '3.11'"]
all = [
    "pyyaml>=5.4",
    "jinja2>=3.0",
    "toml>=0.10; python_version < '3.11'",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/anthropics/doc-guardian"
Documentation = "https://github.com/anthropics/doc-guardian#readme"
Repository = "https://github.com/anthropics/doc-guardian.git"
Issues = "https://github.com/anthropics/doc-guardian/issues"
Changelog = "https://github.com/anthropics/doc-guardian/blob/main/CHANGELOG.md"

[project.scripts]
doc-guardian = "guardian.heal:main"
guardian-heal = "guardian.heal:main"
guardian-install = "guardian.install:main"
guardian-rollback = "guardian.rollback:main"

[tool.setuptools]
packages = ["guardian", "guardian.core", "guardian.healers"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.md", "*.toml", "*.yaml", "*.yml", "*.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py38"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["guardian"]

[tool.coverage.run]
source = ["guardian"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
